+++ include/crab/http/response_parser.hxx
@@ -21,46 +21,46 @@ CRAB_INLINE void ResponseParser::parse(Buffer &buf) {
 // https://www.w3.org/Protocols/rfc2616/rfc2616-sec19.html#sec19.3
 CRAB_INLINE ResponseParser::State ResponseParser::consume(char input) {
 	if (++total_length > max_total_length)
-		throw std::runtime_error("HTTP Header too long - security violation");
+		throw std::runtime_error{"HTTP Header too long - security violation"};
 
 	switch (state) {
 	case HTTP_VERSION_H:
 		if (input != 'H')
-			throw std::runtime_error("Invalid http version, 'H' is expected");
+			throw std::runtime_error{"Invalid http version, 'H' is expected"};
 		return HTTP_VERSION_HT;
 	case HTTP_VERSION_HT:
 		if (input != 'T')
-			throw std::runtime_error("Invalid http version, 'T' is expected");
+			throw std::runtime_error{"Invalid http version, 'T' is expected"};
 		return HTTP_VERSION_HTT;
 	case HTTP_VERSION_HTT:
 		if (input != 'T')
-			throw std::runtime_error("Invalid http version, 'T' is expected");
+			throw std::runtime_error{"Invalid http version, 'T' is expected"};
 		return HTTP_VERSION_HTTP;
 	case HTTP_VERSION_HTTP:
 		if (input != 'P')
-			throw std::runtime_error("Invalid http version, 'P' is expected");
+			throw std::runtime_error{"Invalid http version, 'P' is expected"};
 		return HTTP_VERSION_SLASH;
 	case HTTP_VERSION_SLASH:
 		if (input != '/')
-			throw std::runtime_error("Invalid http version, '/' is expected");
+			throw std::runtime_error{"Invalid http version, '/' is expected"};
 		return HTTP_VERSION_MAJOR_START;
 	case HTTP_VERSION_MAJOR_START:
 		if (!isdigit(input))
-			throw std::runtime_error("Invalid http version major start, must be digit");
+			throw std::runtime_error{"Invalid http version major start, must be digit"};
 		req.http_version_major = input - '0';
 		return HTTP_VERSION_MAJOR;
 	case HTTP_VERSION_MAJOR:
 		if (input == '.')
 			return HTTP_VERSION_MINOR_START;
 		if (!isdigit(input))
-			throw std::runtime_error("Invalid http version major, must be digit");
+			throw std::runtime_error{"Invalid http version major, must be digit"};
 		req.http_version_major = req.http_version_major * 10 + input - '0';
 		if (req.http_version_major > 1)
-			throw std::runtime_error("Unsupported http version");
+			throw std::runtime_error{"Unsupported http version"};
 		return HTTP_VERSION_MAJOR;
 	case HTTP_VERSION_MINOR_START:
 		if (!isdigit(input))
-			throw std::runtime_error("Invalid http version minor start, must be digit");
+			throw std::runtime_error{"Invalid http version minor start, must be digit"};
 		req.http_version_minor = input - '0';
 		return HTTP_VERSION_MINOR;
 	case HTTP_VERSION_MINOR:
@@ -69,26 +69,26 @@ CRAB_INLINE ResponseParser::State ResponseParser::consume(char input) {
 			return STATUS_CODE_1;
 		}
 		if (!isdigit(input))
-			throw std::runtime_error("Invalid http version minor, must be digit");
+			throw std::runtime_error{"Invalid http version minor, must be digit"};
 		req.http_version_minor = req.http_version_minor * 10 + input - '0';
 		if (req.http_version_minor > 99)
-			throw std::runtime_error("Invalid http version minor, too big");
+			throw std::runtime_error{"Invalid http version minor, too big"};
 		return HTTP_VERSION_MINOR;
 	case STATUS_CODE_1:
 		if (is_sp(input))
 			return STATUS_CODE_1;
 		if (!isdigit(input))
-			throw std::runtime_error("Invalid http status code symbol 1, must be digit");
+			throw std::runtime_error{"Invalid http status code symbol 1, must be digit"};
 		req.status = req.status * 10 + input - '0';
 		return STATUS_CODE_2;
 	case STATUS_CODE_2:
 		if (!isdigit(input))
-			throw std::runtime_error("Invalid http status code symbol 2, must be digit");
+			throw std::runtime_error{"Invalid http status code symbol 2, must be digit"};
 		req.status = req.status * 10 + input - '0';
 		return STATUS_CODE_3;
 	case STATUS_CODE_3:
 		if (!isdigit(input))
-			throw std::runtime_error("Invalid http status code symbol 3, must be digit");
+			throw std::runtime_error{"Invalid http status code symbol 3, must be digit"};
 		req.status = req.status * 10 + input - '0';
 		return STATUS_CODE_SPACE;
 	case STATUS_CODE_SPACE:
@@ -98,7 +98,7 @@ CRAB_INLINE ResponseParser::State ResponseParser::consume(char input) {
 		if (input == '\n')
 			return FIRST_HEADER_LINE_START;
 		if (!is_sp(input))
-			throw std::runtime_error("Invalid http status code, space expected");
+			throw std::runtime_error{"Invalid http status code, space expected"};
 		return STATUS_TEXT_START;
 	case STATUS_TEXT_START:
 		if (is_sp(input))
@@ -111,12 +111,12 @@ CRAB_INLINE ResponseParser::State ResponseParser::consume(char input) {
 		if (input == '\n')
 			return FIRST_HEADER_LINE_START;
 		if (is_ctl(input))
-			throw std::runtime_error("Invalid character (control) in status text");
+			throw std::runtime_error{"Invalid character (control) in status text"};
 		req.status_text.push_back(input);
 		return STATUS_TEXT;
 	case STATUS_LINE_LF:
 		if (input != '\n')
-			throw std::runtime_error("Newline is expected");
+			throw std::runtime_error{"Newline is expected"};
 		return FIRST_HEADER_LINE_START;
 	case FIRST_HEADER_LINE_START:  // Cannot contain LWS
 		trim_right(req.status_text);
@@ -126,8 +126,8 @@ CRAB_INLINE ResponseParser::State ResponseParser::consume(char input) {
 		if (input == '\n')
 			return GOOD;
 		if (!is_char(input) || is_ctl(input) || is_tspecial(input))
-			throw std::runtime_error("Invalid character at header line start");
-		header.name.push_back(std::tolower(input));
+			throw std::runtime_error{"Invalid character at header line start"};
+		header.name.push_back(static_cast<char>(std::tolower(input)));
 		return HEADER_NAME;
 	case HEADER_LINE_START:
 		if (is_sp(input)) {
@@ -142,8 +142,8 @@ CRAB_INLINE ResponseParser::State ResponseParser::consume(char input) {
 		if (input == '\n')
 			return GOOD;
 		if (!is_char(input) || is_ctl(input) || is_tspecial(input))
-			throw std::runtime_error("Invalid character at header line start");
-		header.name.push_back(std::tolower(input));
+			throw std::runtime_error{"Invalid character at header line start"};
+		header.name.push_back(static_cast<char>(std::tolower(input)));
 		return HEADER_NAME;
 	case HEADER_NAME:
 		// We relax https://tools.ietf.org/html/rfc7230#section-3.2.4
@@ -151,8 +151,8 @@ CRAB_INLINE ResponseParser::State ResponseParser::consume(char input) {
 			return HEADER_COLON;
 		if (input != ':') {
 			if (!is_char(input) || is_ctl(input) || is_tspecial(input))
-				throw std::runtime_error("Invalid character at header name");
-			header.name.push_back(std::tolower(input));
+				throw std::runtime_error{"Invalid character at header name"};
+			header.name.push_back(static_cast<char>(std::tolower(input)));
 			return HEADER_NAME;
 		}
 		// Next line is understood by GCC
@@ -161,7 +161,7 @@ CRAB_INLINE ResponseParser::State ResponseParser::consume(char input) {
 		if (is_sp(input))
 			return HEADER_COLON;
 		if (input != ':')
-			throw std::runtime_error("':' expected");
+			throw std::runtime_error{"':' expected"};
 		// We will add other comma-separated headers if we need them later
 		header_cms_list = (header.name == string_view{"connection"}) || (header.name == string_view{"transfer-encoding"});
 		return SPACE_BEFORE_HEADER_VALUE;
@@ -177,7 +177,7 @@ CRAB_INLINE ResponseParser::State ResponseParser::consume(char input) {
 		if (input == '\n')
 			return HEADER_LINE_START;
 		if (is_ctl(input))
-			throw std::runtime_error("Invalid character (control) in header value");
+			throw std::runtime_error{"Invalid character (control) in header value"};
 		if (header_cms_list && input == ',') {
 			process_ready_header();
 			header.value.clear();
@@ -187,14 +187,14 @@ CRAB_INLINE ResponseParser::State ResponseParser::consume(char input) {
 		return HEADER_VALUE;
 	case HEADER_LF:
 		if (input != '\n')
-			throw std::runtime_error("Expecting newline");
+			throw std::runtime_error{"Expecting newline"};
 		return HEADER_LINE_START;
 	case FINAL_LF:
 		if (input != '\n')
-			throw std::runtime_error("Expecting final newline");
+			throw std::runtime_error{"Expecting final newline"};
 		return GOOD;
 	default:
-		throw std::logic_error("Invalid response parser state");
+		throw std::logic_error{"Invalid response parser state"};
 	}
 }
 
@@ -206,11 +206,11 @@ CRAB_INLINE void ResponseParser::process_ready_header() {
 	// Those comparisons are by size first so very fast
 	if (header.name == string_view{"content-length"}) {
 		if (req.content_length)
-			throw std::runtime_error("content length specified more than once");
+			throw std::runtime_error{"content length specified more than once"};
 		try {
 			req.content_length = integer_cast<uint64_t>(header.value);
 		} catch (const std::exception &) {
-			std::throw_with_nested(std::runtime_error("Content length is not a number"));
+			std::throw_with_nested(std::runtime_error{"Content length is not a number"});
 		}
 		return;
 	}
@@ -218,7 +218,7 @@ CRAB_INLINE void ResponseParser::process_ready_header() {
 		tolower(header.value);
 		if (header.value == string_view{"chunked"}) {
 			if (!req.transfer_encodings.empty())
-				throw std::runtime_error("chunk encoding must be applied last");
+				throw std::runtime_error{"chunk encoding must be applied last"};
 			req.transfer_encoding_chunked = true;
 			return;
 		}
@@ -246,7 +246,7 @@ CRAB_INLINE void ResponseParser::process_ready_header() {
 			req.connection_upgrade = true;
 			return;
 		}
-		throw std::runtime_error("Invalid 'connection' header value");
+		throw std::runtime_error{"Invalid 'connection' header value"};
 	}
 	if (header.name == string_view{"upgrade"}) {
 		tolower(header.value);
@@ -254,7 +254,7 @@ CRAB_INLINE void ResponseParser::process_ready_header() {
 			req.upgrade_websocket = true;
 			return;
 		}
-		throw std::runtime_error("Invalid 'upgrade' header value");
+		throw std::runtime_error{"Invalid 'upgrade' header value"};
 	}
 	if (header.name == string_view{"sec-websocket-accept"}) {
 		req.sec_websocket_accept = header.value;  // Copy is better here