+++ include/crab/http/server.hxx
@@ -178,7 +178,7 @@ CRAB_INLINE void Server::on_client_handle_request(Client *who, Request &&request
 		return;
 	}
 	if (who->get_state() == ServerConnection::RESPONSE_HEADER && !who->d_handler)
-		throw std::logic_error("r_handler must either write response, call postpone_response or web_socket_upgrade");
+		throw std::logic_error{"r_handler must either write response, call postpone_response or web_socket_upgrade"};
 }
 
 CRAB_INLINE void Server::on_client_handle_message(Client *who, WebMessage &&message) {