+++ include/crab/network.hxx
@@ -20,30 +20,31 @@ CRAB_INLINE void PerformanceStats::push_record(const char *event_type_literal, i
 }
 
 CRAB_INLINE void PerformanceStats::print_records(std::ostream &out) {
-	for (const auto &p : get_records()) {
+	for (const auto &p : performance) {
 		auto mksec = std::chrono::duration_cast<std::chrono::microseconds>(p.tm.time_since_epoch()).count();
 		auto sec   = mksec / 1000000;
 		out << "* " << sec << "." << mksec % 1000000 << " " << p.event_type << " " << p.count << std::endl;
 	}
+	out << "---- total records " << performance.size() << std::endl;
 	clear_records();
 }
 
 CRAB_INLINE Address::Address(const std::string &ip, uint16_t port) {
 	if (!parse(*this, ip, port))
 		if (!parse(*this, ip, port))
-			throw std::runtime_error("Address failed to parse, numeric_host='" + ip + "'");
+			throw std::runtime_error{"Address failed to parse, numeric_host='" + ip + "'"};
 }
 
 CRAB_INLINE Address::Address(const std::string &ip_port) {
 	if (!parse(*this, ip_port))
-		throw std::runtime_error("Address failed to parse, must be <ip>:<port> numeric_host_port='" + ip_port + "'");
+		throw std::runtime_error{"Address failed to parse, must be <ip>:<port> numeric_host_port='" + ip_port + "'"};
 }
 
 CRAB_INLINE bool Address::parse(Address &address, const std::string &ip_port) {
 	size_t pos = ip_port.find(':');
 	if (pos == std::string::npos)
 		return false;
-	uint16_t port = integer_cast<uint16_t>(ip_port.substr(pos + 1));
+	auto port = integer_cast<uint16_t>(ip_port.substr(pos + 1));
 	return parse(address, ip_port.substr(0, pos), port);
 }
 
@@ -176,7 +177,7 @@ CRAB_INLINE void RunLoop::run() {
 	}
 }
 
-CRAB_INLINE steady_clock::time_point RunLoop::now() { return links.now; }
+CRAB_INLINE steady_clock::time_point RunLoop::now() const { return links.now; }
 
 CRAB_INLINE Thread::Thread(std::function<void()> &&fun)
 #if __cplusplus >= 201402L
@@ -398,7 +399,7 @@ CRAB_INLINE Address DNSResolver::sync_resolve_single(const std::string &host_nam
 	arr = sync_resolve(host_name, port, false, true);
 	if (!arr.empty())
 		return arr.front();
-	throw std::runtime_error("Failed to resolve host '" + host_name + "'");
+	throw std::runtime_error{"Failed to resolve host '" + host_name + "'"};
 }
 
 }  // namespace crab