+++ include/crab/streams.hpp
@@ -72,9 +72,9 @@ public:
 	using OStream::write_some;  // Version for other char types
 
 	void clear() { read_pos = write_pos = 0; }
-	void clear(size_t newsize) {
+	void clear(size_t new_size) {
 		clear();
-		impl.resize(newsize);
+		impl.resize(new_size);
 	}
 
 	bool full() const { return read_pos + impl.size() == write_pos; }